fix: explicitly cast calloc_a variadic length arguments to size_t
On 64-bit architectures (e.g., aarch64), `size_t` is 8 bytes, while `int`
remains 4 bytes. The `calloc_a` macro uses variadic arguments (`...`) to
read memory allocation lengths via `va_arg(ap, size_t)`.

If a 32-bit integer (like `tlen`, `dlen`, `txt_len`, or `0` in a ternary
operator) is passed without an explicit cast, the compiler will not
automatically promote it to 64-bit. As a result, `va_arg` reads 8 bytes,
consuming the 32-bit value plus adjacent stack garbage. This triggers
massive, uncontrolled memory allocations (e.g., attempting to malloc
508GB) and causes immediate OOM crashes.

This patch explicitly casts all length arguments passed to `calloc_a`
to `(size_t)` in `cache.c` and `service.c`, ensuring safe memory
allocation across all architectures.

+++ service.c
@@ -259,11 +259,11 @@ service_load_blob(struct blob_attr *b)
 
 	n = strlen(blobmsg_name(b));
 	s = calloc_a(sizeof(*s),
-		&d_id, n + 1,
-		&d_hostname, _tb[SERVICE_HOSTNAME] ? strlen(blobmsg_get_string(_tb[SERVICE_HOSTNAME])) + 1 : 0,
-		&d_instance, _tb[SERVICE_INSTANCE] ? strlen(blobmsg_get_string(_tb[SERVICE_INSTANCE])) + 1 : 0,
+		&d_id, (size_t)(n + 1),
+		&d_hostname, _tb[SERVICE_HOSTNAME] ? strlen(blobmsg_get_string(_tb[SERVICE_HOSTNAME])) + 1 : (size_t)0,
+		&d_instance, _tb[SERVICE_INSTANCE] ? strlen(blobmsg_get_string(_tb[SERVICE_INSTANCE])) + 1 : (size_t)0,
 		&d_service, strlen(blobmsg_get_string(_tb[SERVICE_SERVICE])) + 1,
-		&d_txt, txt_len);
+		&d_txt, (size_t)txt_len);
 	if (!s)
 		return;
